Fixed some compilation warnings, trying to work with VSCode on Windows

+++ src/PipelineManager.cpp
@@ -237,7 +237,7 @@ VkPipelineColorBlendStateCreateInfo PipelineManager::GetBlendState(size_t numAtt
         VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_STATE_CREATE_INFO;
     colorBlending.logicOpEnable = VK_FALSE;
     colorBlending.logicOp = VK_LOGIC_OP_COPY;  // Optional
-    colorBlending.attachmentCount = numAttachments;
+    colorBlending.attachmentCount = static_cast<uint32_t>(numAttachments);
     colorBlending.pAttachments = m_aBlendModes.data();
     colorBlending.blendConstants[0] = 0.0f;  // Optional
     colorBlending.blendConstants[1] = 0.0f;  // Optional
